+++ stats-service/Dockerfile
