+++ stats-service/client/src/main/ru.practicum/java/client/StatClient.java
@@ -1,12 +1,12 @@
 package client;
 
-import dto.HitDto;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.web.util.DefaultUriBuilderFactory;
+import ru.practicum.dto.HitDto;
 
 import java.util.List;
 import java.util.Map;
