+++ stats-service/dto/pom.xml
@@ -23,4 +23,5 @@
             <artifactId>lombok</artifactId>
         </dependency>
     </dependencies>
+
 </project>
