+++ stats-service/server/src/main/java/ru/practicum/server/StatServiceApplication.java
@@ -1,4 +1,4 @@
-package server;
+package ru.practicum.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
