+++ stats-service/server/src/main/java/ru/practicum/server/exeption/NotSupportedException.java
@@ -1,4 +1,4 @@
-package server.exeption;
+package ru.practicum.server.exeption;
 
 public class NotSupportedException extends RuntimeException {
     public NotSupportedException(String message) {
