+++ stats-service/server/src/main/java/ru/practicum/server/model/Hit.java
@@ -1,4 +1,4 @@
-package server.model;
+package ru.practicum.server.model;
 
 import lombok.*;
