+++ stats-service/server/src/main/java/ru/practicum/server/repository/StatsRepository.java
@@ -1,23 +1,23 @@
-package server.repository;
+package ru.practicum.server.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import server.model.Hit;
-import server.model.OutStats;
+import ru.practicum.server.model.Hit;
+import ru.practicum.server.model.OutStats;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface StatsRepository extends JpaRepository<Hit, Long> {
 
-    @Query("select new server.model.OutStats(h.app, h.uri, count(h.ip))" +
+    @Query("select new OutStats(h.app, h.uri, count(h.ip))" +
             "from Hit h " +
             "where h.timestamp between ?1 and ?2 " +
             "group by h.app, h.uri " +
             "order by count(h.ip) desc")
     List<OutStats> findAllWithoutUris(LocalDateTime start, LocalDateTime end);
 
-    @Query("select new server.model.OutStats(h.app, h.uri, count(h.ip))" +
+    @Query("select new OutStats(h.app, h.uri, count(h.ip))" +
             "from Hit h " +
             "where h.timestamp between ?1 and ?2 " +
             "and h.uri in (?3) " +
@@ -25,7 +25,7 @@ public interface StatsRepository extends JpaRepository<Hit, Long> {
             "order by count(h.ip) desc")
     List<OutStats> findAllNotUnique(LocalDateTime start, LocalDateTime end, List<String> uris);
 
-    @Query("select new server.model.OutStats(h.app, h.uri, count(distinct h.ip))" +
+    @Query("select new OutStats(h.app, h.uri, count(distinct h.ip))" +
             "from Hit h " +
             "where h.timestamp between ?1 and ?2 " +
             "and h.uri in (?3) " +
